+++ CHANGELOG.rst
@@ -3,3 +3,9 @@
     * stack_operations created
     * stack_operations include get_max, get_min . get_avg and print_stack
     * 12 test cases added for stack adt
+
+0.1.1
+    * Linked List ADT has been added to the package
+    * Implementation of Node for singly list
+    * Implementation of Node for doubly list
+    * Stack Ops method included , how many spaces left